StatusEvent and Long IDs

+++ app/src/main/kotlin/org/cirruslabs/sq/github/model/Issue.kt
@@ -1,7 +1,7 @@
 package org.cirruslabs.sq.github.model
 
 class Issue {
-  var id: Int = 0
-  var number: Int = 0
+  var id: Long = 0
+  var number: Long = 0
   var labels: List<String> = emptyList()
 }
