+++ app/src/main/kotlin/org/cirruslabs/sq/github/model/Label.kt
@@ -1,6 +1,6 @@
 package org.cirruslabs.sq.github.model
 
 class Label {
-  var id: Int = 0
+  var id: Long = 0
   var name: String = ""
 }
